scripts/requirements.txt
databento>=0.41.0
supabase>=2.0.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
autogluon.tabular>=1.2.0
arch>=7.0.0
